rename: 유저 사용량 측정 컴포넌트 이름 변경
